third_party/WebKit/LayoutTests/external/wpt/hr-time/timeOrigin_window_worker.html

<!doctype html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>
async_test(function(t) {
  const windowOrigin = performance.timeOrigin;
  const workerScript = 'postMessage({timeOrigin: performance.timeOrigin})';
  const blob = new Blob([workerScript]);
  const worker = new Worker(URL.createObjectURL(blob));
  worker.addEventListener('message', t.step_func_done(function(event) {
    const workerOrigin = event.data.timeOrigin;
    assert_greater_than(workerOrigin, windowOrigin, 'The worker timeOrigin should be greater than the window timeOrigin.');
    assert_less_than(workerOrigin - windowOrigin, 500, 'Difference between window and worker timeOrigins should be small.');
  }));
}, 'Window and worker timeOrigins are close when created one after another.');
</script>
</body>
</html>
